Users update

+++ Es/Components/Users/users.controller.js
@@ -1,8 +1,18 @@
 import * as usersService from "./users.service.js";
 
-export const getUserById = (req, res) => {
+export const getUserById = async (req, res) => {
+ const schema = z.object({
+    params: z.object({
+      id: z.preprocess((val) => Number(val), z.number().positive()),
+    }),
+  });
+    const isValidData = await schema.safeParseAsync({
+      params: req.params,
+    });
+    if (!isValidData.success) {
+        throw new ErrorWithStatus(422, isValidData.error.issues);
+    }
   const user = usersService.getUserById(Number(req.params.id));
-
   res.status(200).json(user);
 };
 
@@ -12,22 +22,64 @@ export const getAllUsers = (req, res) => {
   res.status(200).json(users);
 };
 
-export const createUser = (req, res) => {
+export const createUser = async (req, res) => {
+    const schema = z.object({
+    params: z.object({
+      name: z.string(),
+      email: z.string(),
+      age: z.number().int().positive(),
+      isActive: z.boolean(),
+    }),
+  });
+    const isValidData = await schema.safeParseAsync({
+      params: req.params,
+    });
+    if (!isValidData.success) {
+        throw new ErrorWithStatus(422, isValidData.error.issues);
+    }
   const user = usersService.createUser(req.body);
 
   res.status(201).json(user);
 };
 
-export const updateUser = (req, res) => {
+export const updateUser = async (req, res) => {
+  const schema = z.object({
+    params: z.object({
+      id: z.preprocess((val) => Number(val), z.number().positive()),
+    }),
+    body: z.object({
+      name: z.string(),
+      email: z.string(),
+      age: z.number().int().positive(),
+      isActive: z.boolean(),
+    }),
+  });
+  const isValidData = await schema.safeParseAsync({
+    params: req.params,
+    body: req.body,
+  });
   const user = usersService.updateUser({
     ...req.body,
     id: Number(req.params.id),
   });
-
+    if (!isValidData.success) {
+      throw new ErrorWithStatus(422, isValidData.error.issues);
+    }
   res.status(200).json(user);
 };
 
-export const deleteUser = (req, res) => {
+export const deleteUser = async (req, res) => {
+  const schema = z.object({
+    params: z.object({
+      id: z.preprocess((val) => Number(val), z.number().positive()),
+    }),
+  });
+  const isValidData = await schema.safeParseAsync({
+    params: req.params,
+  });
+  if (!isValidData.success) {
+    throw new ErrorWithStatus(422, isValidData.error.issues);
+  }
   const result = usersService.deleteUser(Number(req.params.id));
 
   res.status(200).json(result);
